refactor(WooCommerce::Spider): add parent class
