+++ spiders/cartonazo_spider.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 # Cartonazo store spider
-# Engine: woocommerce
-class CartonazoSpider < ApplicationSpider
+class CartonazoSpider < EcommerceEngines::WooCommerce::Spider
   @name = "cartonazo_spider"
   @store = {
     name: "Cartonazo",
@@ -10,67 +9,4 @@ class CartonazoSpider < ApplicationSpider
   }
   @start_urls = ["https://cartonazo.com/categoria-producto/juego-de-mesa"]
   @config = {}
-
-  def parse(response, url:, data: {})
-    items = parse_index(response, url:)
-    items.each { |item| send_item item }
-
-    paginate(response, url)
-  end
-
-  def parse_index(response, url:, data: {})
-    listings = response.css("ul.products li.product")
-    listings.map { |listing| parse_product_node(listing, url:) }
-  end
-
-  def parse_product_node(node, url:)
-    {
-      url: get_url(node),
-      title: get_title(node),
-      price: get_price(node),
-      stock: purchasable?(node),
-      image_url: get_image_url(node)
-    }
-  end
-
-  def next_page_url(response, url)
-    next_page = response.at_css("nav.woocommerce-pagination li a.next")
-    return unless next_page
-
-    absolute_url(next_page[:href], base: url)
-  end
-
-  private
-
-  def paginate(response, url)
-    next_page_url = next_page_url(response, url)
-    request_to(:parse, url: next_page_url) if next_page_url
-  end
-
-  def get_url(node)
-    node.at_css("a")[:href]
-  end
-
-  def get_title(node)
-    node.at_css("h2").text
-  end
-
-  def get_price(node)
-    price_node = node.css("span.price bdi").last
-    scan_int(price_node.text) if price_node
-  end
-
-  def in_stock?(node)
-    node.classes.include?("instock")
-  end
-
-  def purchasable?(node)
-    in_stock?(node)
-  end
-
-  def get_image_url(node)
-    node.at_css("img")["srcset"].split(",").last.split.first
-  rescue NoMethodError
-    nil
-  end
 end
